pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ansys-tools-repo-sync"
authors = ["ANSYS, Inc."]
maintainers = ["PyAnsys developers <pyansys.support@ansys.com>"]
readme = "README.rst"
homepage = "https://github.com/ansys/ansys-tools-repo-sync"
license = "MIT"
classifiers = ["License :: OSI Approved :: MIT License"]
version = "0.1.16"
description = "Synchronize the content of two different repositories."


packages = [
    { include = "ansys"},
]

[tool.poetry.dependencies]
python = "^3.7"
importlib-metadata = {version = "^4.0"}

pre-commit = "^2.15.0"
pygithub = "1.55"
click = "^8.0.4"

Sphinx =  {version=">=4.4,<6.0", optional=true}
numpydoc =  {version="^1.2", optional=true}
pyansys_sphinx_theme =  {version="^0.2", optional=true}
sphinx-copybutton =  {version=">=0.4,<0.6", optional=true}

codespell = {version="^2.1", optional=true}
flake8 = {version="^3.9", optional=true}

pytest = {version="^7.0", optional=true}
pytest-cov = {version="^3.0", optional=true}

[tool.codespell]
skip="*.pyc,*.xml,*.txt,*.gif,*.png,*.jpg,*.js,*.html,*.doctree,*.ttf,*.woff,*.woff2,*.eot,*.mp4,*.inv,*.pickle,*.ipynb,flycheck*,./.git/*,./.hypothesis/*,*.yml,./docs/build/*,./docs/images/*,./dist/*,*~,.hypothesis*,./docs/source/examples/*,*cover,*.dat,*.mac"


[tool.black]
line-length = 120
target-version = ['py37']

[tool.pydocstyle]
convention = "numpy"

[tool.mypy]
strict = true

ignore_missing_imports = true


[tool.poetry.extras]
docs = ["Sphinx", "numpydoc", "pyansys_sphinx_theme", "sphinx-copybutton"]
style = ["codespell", "flake8"]
test = ["pytest", "pytest-cov"]

[tool.poetry.scripts]
repo-sync = "ansys.tools.repo_sync._cli:synchronize"
